refactor(): add input handlers
- move input type handling responsibility to external modules
- add common interface for input handling objects
