+++ src/inputHandlers/IMediaInfoInputHandler.ts
@@ -1,0 +1,6 @@
+import { ReadStream } from 'fs';
+import { Duplex } from 'stream';
+
+export interface IMediaInfoInputHandler {
+  openStream(input: string): ReadStream | Duplex
+}
